B4491 [语言月赛 202602] 数学补习

+++ B4082 [语言月赛 202412] 古希腊掌管罚时的神.cpp
@@ -4,7 +4,7 @@
 #include<iostream>
 using namespace std;
 int main(){
-    int n,time=0;
+    long long n,time=0;
     cin>>n;
     for(int i=1;i<=n;i++){
         int a,b;
